(ubw) fixed some bugs

+++ src/library/stdio.c
@@ -67,7 +67,7 @@ int printf(const char *format, ...) {
                     break;
                 }
                 case 'c': {
-                    PutByteUART((uint8_t)*arg);
+                    PutByteUART((int)*arg);
                     ++arg;
                     ++i;
                     break;
